docs : product와 search service 나눔.
