+++ ducktem-api/src/main/java/com/ducktem/ducktemapi/service/ProductService.java
@@ -12,22 +12,7 @@ import com.ducktem.ducktemapi.entity.Product;
 
 public interface ProductService {
 	ProductDetailResponse get(Long id);
-
 	Product add(Product product, String regMemberId);
-
-	List<ProductPreviewResponse> getList(Pageable pageable);
-
-	List<ProductPreviewResponse> getListBySearch(Pageable pageable, String query);
-
-	Long getCountByQuery(String query, Integer[] category);
-
-	List<ProductPreviewResponse> getListByCategory(Pageable pageable, Integer[] category);
-
-	List<ProductPreviewResponse> getListByCategoryAndSearch(Pageable pageable,String query,Integer[] category);
-	//높은가격순
-	List<ProductPreviewResponse> getListByCategoryAndSearchOrderByPriceDesc(Pageable pageable,String query,Integer[] category);
-	//낮은가격순
-	List<ProductPreviewResponse> getListByCategoryAndSearchOrderByPrice(Pageable pageable,String query,Integer[] category);
 
 
 }
