+++ ducktem-api/src/main/java/com/ducktem/ducktemapi/service/ProductServiceImpl.java
@@ -37,86 +37,6 @@ public class ProductServiceImpl implements ProductService {
 		return ProductDetailResponse.from(product);
 	}
 
-	// 카테고리와 검색어로 검색한 결과를 Count. Long 타입으로 반환.
-	@Override
-	public Long getCountByQuery(String query, Integer[] categoryId) {
-		List<Category> category = (List<Category>)categoryRepository.findByIdIn(categoryId);
-		String newQuery = "%"+query+"%";
-
-		return productRepository.countProductByNameLikeQuery(newQuery,category);
-	}
-	// 카테고리로 검색한 상품 결과를 상품 레포지토리에서 paging 이후 productPreview로 바꿔서 반환.
-	@Override
-	public List<ProductPreviewResponse> getListByCategory(Pageable pageable, Integer[] categoryId) {
-
-
-		 List<Category> category = (List<Category>)categoryRepository.findByIdIn(categoryId);
-
-		return productRepository.findByCategoryInOrderByIdDesc(pageable, category)
-								.map(ProductPreviewResponse::from)
-								.toList();
-	}
-	// 카테고리와 검색어로 검색한 상품 결과를 상품 레포지토리에서 paging 이후 productPreview로 바꿔서 반환.
-	@Override
-	public List<ProductPreviewResponse> getListByCategoryAndSearch(
-					Pageable pageable, String query, Integer[] categoryId) {
-
-		List<Category> category = (List<Category>)categoryRepository.findByIdIn(categoryId);
-		String newQuery = "%"+query+"%";
-		return productRepository.findByNameQueryAndCategoryInOrderByIdDesc(pageable,newQuery, category)
-								.map(ProductPreviewResponse::from)
-								.toList();
-	}
-
-
-
-	// 상품 레포지토리에서 paging 이후 productPreview로 바꿔서 반환.
-	@Override
-	@Transactional
-	public List<ProductPreviewResponse> getList(Pageable pageable) {
-
-		return productRepository.findAll(pageable)
-			.map(ProductPreviewResponse::from)
-			.toList();
-	}
-
-	@Override
-	public List<ProductPreviewResponse> getListBySearch(Pageable pageable, String query) {
-
-		String newQuery = "%"+query+"%";
-
-		return productRepository.findByNameQueryOrderByIdDesc(pageable, newQuery)
-			.map(ProductPreviewResponse::from)
-			.toList();
-	}
-
-
-	//---정렬옵션
-	//높은가격순
-	@Override
-	public List<ProductPreviewResponse> getListByCategoryAndSearchOrderByPriceDesc(Pageable pageable, String query,
-		Integer[] categoryId) {
-		List<Category> category = (List<Category>)categoryRepository.findByIdIn(categoryId);
-		String newQuery = "%"+query+"%";
-		return productRepository.findByNameQueryAndCategoryInOrderByPriceDesc(pageable,newQuery, category)
-			.map(ProductPreviewResponse::from)
-			.toList();
-	}
-	//낮은가격순
-	@Override
-	public List<ProductPreviewResponse> getListByCategoryAndSearchOrderByPrice(Pageable pageable, String query,
-		Integer[] categoryId) {
-		List<Category> category = (List<Category>)categoryRepository.findByIdIn(categoryId);
-		String newQuery = "%"+query+"%";
-		return productRepository.findByNameQueryAndCategoryInOrderByPrice(pageable,newQuery, category)
-			.map(ProductPreviewResponse::from)
-			.toList();
-	}
-
-
-
-
-
 	@Override
 	@Transactional
 	public Product add(Product product, String regMemberId) {
